admin: add and move code in team route
